+++ packages/datepickers/examples/compact.md
